+++ modules/core/src/client.ts
@@ -1,8 +1,8 @@
 /**
  * @module core
  */
-import { Account, GameState, isGameState } from './models';
-import { Socket } from 'net';
+import { Account, GameState, isGameState, Server } from './models';
+import { Socket, createConnection } from 'net';
 import {
   PacketIO, HelloPacket, PacketType,
   UpdateAckPacket, GotoPacket, GotoAckPacket,
@@ -15,6 +15,8 @@ import {
 import { Logger, LogLevel, Classes, PlayerData } from '@n2/common';
 import { StatusParser } from './parsers';
 import { MapInfo } from './models/map';
+import { SocksClient } from 'socks';
+import { Proxy } from './models/proxy';
 
 /**
  * A lightweight client for a clientless application
@@ -50,6 +52,10 @@ export class Client {
    */
   account: Account;
   /**
+   * A proxy to route game traffic through.
+   */
+  proxy: Proxy;
+  /**
    * @readonly
    * Whether or not the socket is currently connected.
    */
@@ -68,8 +74,16 @@ export class Client {
     return Date.now() - this._time;
   }
 
+  /**
+   * The last server that this client was connected to.
+   */
+  get lastServer(): Server {
+    return this._lastServer;
+  }
+
   private _connected: boolean;
   private _time: number;
+  private _lastServer: Server;
 
   private eventHandlers: Map<string, (...args: any[]) => void>;
 
@@ -77,16 +91,18 @@ export class Client {
    * Creates a new client for the specified account.
    * @param account The account to create this client for.
    */
-  constructor(account: Account, socket?: Socket) {
-    if (typeof account.guid !== 'string' && typeof account.password !== 'string') {
+  constructor(account: Account) {
+    if (!account || typeof account.guid !== 'string' && typeof account.password !== 'string') {
       throw new TypeError(`Parameter "account" must be an Account, not ${typeof account}`);
     }
-    if (socket && !(socket instanceof Socket)) {
-      throw new TypeError(`Parameter "socket" must be a Socket, not ${typeof socket}`);
-    }
     this.account = account;
+    if (!this.account.name) {
+      let crypto = require('crypto');
+      this.account.name = `Client ${crypto.randomBytes(2).toString('hex')}`;
+      crypto = null;
+    }
     this._connected = false;
-    this.io = new PacketIO(socket || new Socket());
+    this.io = new PacketIO();
     this._time = Date.now();
     this.eventHandlers = new Map([
       ['connect', this.onConnect.bind(this)],
@@ -94,14 +110,10 @@ export class Client {
       ['error', this.onError.bind(this)],
     ]);
 
-    for (const kvp of this.eventHandlers) {
-      this.io.socket.on(kvp[0], kvp[1]);
-    }
-
     this.io
       .on(PacketType.CREATE_SUCCESS, (createSuccess: CreateSuccessPacket) => {
         this.objectId = createSuccess.objectId;
-        Logger.log('Client', 'Client connected.', LogLevel.Success);
+        this.log('Client connected.', LogLevel.Success);
       })
       .on(PacketType.MAPINFO, (mapInfo: MapInfoPacket) => {
         this.map = {
@@ -161,21 +173,56 @@ export class Client {
   }
 
   /**
+   * Attaches the client to the `Socket`.
+   * @param socket The socket to attach to.
+   */
+  attach(socket: Socket): void {
+    if (!(socket instanceof Socket)) {
+      throw new TypeError(`Parameter "socket" should be a Socket, not ${typeof socket}`);
+    }
+    if (this.io.socket) {
+      this.detach();
+    }
+    this.io.attach(socket);
+    for (const kvp of this.eventHandlers) {
+      this.io.socket.on(kvp[0], kvp[1]);
+    }
+  }
+
+  /**
+   * Detaches this client from its socket.
+   */
+  detach(): void {
+    if (this.io.socket) {
+      for (const kvp of this.eventHandlers) {
+        this.io.socket.removeListener(kvp[0], kvp[1]);
+      }
+      this.io.socket.destroy();
+      this.io.detach();
+    }
+  }
+
+  /**
    * Connects to the specified IP with the given state.
    * @param ip The IP to connect to.
    * @param gameState The state to use when connecting.
    */
-  connect(ip: string, gameState: GameState): Promise<void> {
-    if (typeof ip !== 'string') {
-      throw new TypeError(`Parameter "ip" must be a string, not ${typeof ip}`);
+  connect(server: Server, gameState: GameState): Promise<void> {
+    if (!server || typeof server.ip !== 'string') {
+      throw new TypeError(`Parameter "server" must be a Server, not ${typeof server}`);
     }
     if (!isGameState(gameState)) {
       throw new TypeError(`Parameter "gameState" must be a GameState, not ${typeof gameState}`);
     }
-    if (this._connected) {
-      this.io.socket.destroy();
+
+    if (this.io.socket) {
+      this.detach();
     }
-    this.io.socket.connect(2050, ip, () => {
+
+    return this.getConnection(server.ip).then((socket) => {
+      this._lastServer = server;
+      this.onConnect();
+      this.attach(socket);
       const helloPacket: HelloPacket = new HelloPacket();
       helloPacket.buildVersion = gameState.buildVersion;
       helloPacket.gameId = gameState.gameId;
@@ -199,9 +246,9 @@ export class Client {
           this.io.send(create);
         }
       });
-    });
-    return new Promise((resolve) => {
-      this.io.once(PacketType.CREATE_SUCCESS, () => resolve());
+      return new Promise<void>((resolve) => {
+        this.io.once(PacketType.CREATE_SUCCESS, () => resolve());
+      });
     });
   }
 
@@ -219,31 +266,67 @@ export class Client {
     }
   }
 
-  /**
-   * Destroys the resources held by this client.
-   */
-  destroy() {
-    this.io.destroy();
-    for (const kvp of this.eventHandlers) {
-      this.io.socket.removeListener(kvp[0], kvp[1]);
+  private getConnection(ip: string): Promise<Socket> {
+    if (this.proxy) {
+      return SocksClient.createConnection({
+        proxy: {
+          ipaddress: this.proxy.host,
+          port: this.proxy.port,
+          type: this.proxy.type,
+          userId: this.proxy.userId,
+          password: this.proxy.password
+        },
+        command: 'connect',
+        destination: {
+          host: ip,
+          port: 2050
+        }
+      }).then((established) => {
+        this.log('Proxy connected.', LogLevel.Success);
+        return established.socket;
+      });
+    } else {
+      return new Promise((resolve, reject) => {
+        const socket = createConnection(2050, ip, () => {
+          socket.removeAllListeners('error');
+          resolve(socket);
+        }).once('error', reject);
+      });
     }
-    this.eventHandlers.clear();
-    this.io.socket.destroy();
   }
 
   private onConnect() {
+    if (!this._lastServer) {
+      this._lastServer = {
+        name: 'Unknown server.',
+        ip: this.io.socket.remoteAddress,
+        usage: 0,
+        location: {
+          lat: 0,
+          long: 0
+        }
+      };
+    }
+    this.log(`Socket connected to ${this.lastServer.name || this.lastServer.ip}.`, LogLevel.Success);
     this._connected = true;
-    Logger.log('Client', 'Socket connected.', LogLevel.Success);
   }
 
   private onClose() {
     this._connected = false;
     this.position = undefined;
     this.objectId = undefined;
-    Logger.log('Client', 'Socket closed.', LogLevel.Warning);
+    this.log('Socket closed.', LogLevel.Warning);
   }
 
   private onError(error: Error) {
-    Logger.log('Client', `Socket error: ${error.message}`, LogLevel.Error);
+    this.log(`Socket error: ${error.message}`, LogLevel.Error);
+  }
+
+  private log(msg: string, level: LogLevel = LogLevel.Message): void {
+    let name = 'Client';
+    if (this.account && this.account.name) {
+      name = this.account.name;
+    }
+    Logger.log(name, msg, level);
   }
 }