+++ modules/core/changelog.md
@@ -2,6 +2,16 @@
 
 This document outlines all notable changes to this project.
 
+## [0.2.0] - 2018-10-02
+### Added
++ A dependency on the `socks` npm module.
++ Proxy support for the `Client` class. If the `proxy` property is a proxy object, `connect` will route traffic through it.
++ The `lastServer` property to the Client class.
++ The optional `name` property to the `Account` interface. This is used for client logging.
+
+### Removed
++ Optional `Socket` parameter from `Client` constructor.
+
 ## [0.1.2] - 2018-10-02
 ### Changed
 + Email and password are encrypted in the `connect` method